Wrap Criteria into a Grid item
